refactor: 발표 분석 로직 구조 개선 및 데이터 레이어 통합
* **refactor: 발표 분석 요청 파라미터 구조 변경 및 UseCase 고도화**
    * `AnalyzePresentationUseCase` 및 `PresentationRepository`에서 `PresentationAnalysisRequest` 객체 대신 개별 필드(name, date, category 등)를 직접 전달받도록 변경했습니다.
    * `AnalyzePresentationRecordingUseCase`를 삭제하고 관련 기능을 `AnalyzePresentationUseCase`로 통합했습니다.

* **refactor: Practice 모듈 내 발표 분석 로직 제거 및 Presentation 모듈로 일원화**
    * `PracticeRemoteDataSource`, `PracticeRepository` 및 관련 구현체에서 발표 분석 관련 메서드(`analyzePresentationRecording`)를 삭제했습니다.
    * `PracticeMapper`에서 발표 분석 응답 변환 로직을 제거하고, `RecordingSpeed` 및 `PracticeRecordingOverallEvaluation`의 enum factory 메서드를 사용하도록 개선했습니다.

* **refactor: PresentationRemoteDataSource 파일 업로드 및 매핑 로직 개선**
    * `analyzePresentation` 호출 시 `readBytes()` 대신 `ChannelProvider`와 `toByteReadChannel()`을 사용한 스트리밍 방식으로 멀티파트 데이터를 처리하도록 개선하여 메모리 효율성을 높였습니다.
    * 대본 파일 업로드를 위한 `scriptFile` 파트 처리 로직(`appendTextPart`)을 추가했습니다.
    * `PresentationMapper` 내 수동 매핑 로직을 도메인 모델의 `from()` 메서드 호출 방식으로 전환하여 응집도를 높였습니다.

* **cleanup: 불필요한 클래스 및 테스트 코드 제거**
    * 더 이상 사용되지 않는 `PresentationAnalysisRequest` 데이터 모델을 제거했습니다.
    * 구조 변경에 따라 유효하지 않게 된 `PresentationRepositoryImplTest`를 삭제했습니다.
    * `AppErrorExt` 내 불필요한 들여쓰기를 정리했습니다.
