+++ Prezel/core/data/src/main/java/com/team/prezel/core/data/mapper/PresentationMapper.kt
@@ -1,9 +1,9 @@
 package com.team.prezel.core.data.mapper
 
+import com.team.prezel.core.model.practice.RecordingSpeed
 import com.team.prezel.core.model.presentation.Audience
 import com.team.prezel.core.model.presentation.Category
 import com.team.prezel.core.model.presentation.ExpectedQuestion
-import com.team.prezel.core.model.presentation.PresentationAnalysisRequest
 import com.team.prezel.core.model.presentation.PresentationAnalysisSummary
 import com.team.prezel.core.model.presentation.PresentationGrowthPoint
 import com.team.prezel.core.model.presentation.PresentationScriptDetail
@@ -12,7 +12,6 @@ import com.team.prezel.core.model.presentation.Purpose
 import com.team.prezel.core.model.presentation.ScriptCorrection
 import com.team.prezel.core.model.presentation.Style
 import com.team.prezel.core.model.presentation.WordAnalysisDetail
-import com.team.prezel.core.network.model.presentation.PresentationAnalysisRequestParts
 import com.team.prezel.core.network.model.presentation.PresentationExpectedQuestionResponse
 import com.team.prezel.core.network.model.presentation.PresentationGrowthResponse
 import com.team.prezel.core.network.model.presentation.PresentationScriptAnalysisResponse
@@ -20,37 +19,21 @@ import com.team.prezel.core.network.model.presentation.PresentationScriptDetailR
 import com.team.prezel.core.network.model.presentation.PresentationSummaryResponse
 import com.team.prezel.core.network.model.presentation.PresentationWordAnalysisResponse
 import com.team.prezel.core.network.model.presentation.PresentationWordDetailResponse
-import kotlinx.datetime.LocalDateTime
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toInstant
-import kotlin.time.Instant
-
-internal fun PresentationAnalysisRequest.toRequestParts(): PresentationAnalysisRequestParts =
-    PresentationAnalysisRequestParts(
-        name = title,
-        date = dateTime.toString(),
-        type = category.toServerType(),
-        purpose = purpose.toServerPurpose(),
-        style = style.toServerStyle(),
-        audience = audience.toServerAudience(),
-        script = script,
-        audioFilePath = audioFilePath,
-    )
 
 internal fun PresentationSummaryResponse.toDomain(): PresentationAnalysisSummary =
     PresentationAnalysisSummary(
         presentationId = presentationId,
         analysisResultId = analysisResultId,
         title = name,
-        category = type.toDomainCategory(),
-        purpose = purpose.toDomainPurpose(),
-        style = style.toDomainStyle(),
-        audience = audience.toDomainAudience(),
-        analyzedAt = analysisDate.toAnalysisInstant(),
+        category = Category.from(value = type),
+        purpose = Purpose.from(value = purpose),
+        style = Style.from(value = style),
+        audience = Audience.from(value = audience),
+        analyzedAt = analysisDate,
         durationSeconds = durationSeconds,
         formattedDuration = formattedDuration,
         spm = spm,
-        speedEvaluation = speedEval,
+        speedEvaluation = RecordingSpeed.from(value = speedEval),
         summaryFeedback = summaryFeedback,
         accuracyScore = accuracyScore,
         scriptMatchRate = scriptMatchRate,
@@ -107,75 +90,3 @@ internal fun PresentationWordAnalysisResponse.toDomain(): WordAnalysisDetail =
         startTimeMs = startTimeMs,
         endTimeMs = endTimeMs,
     )
-
-private fun Category.toServerType(): String =
-    when (this) {
-        Category.PERSUASION -> "OFFER"
-        Category.EVENT -> "EVENT"
-        Category.EDUCATION -> "EDUCATION"
-        Category.REPORT -> "WORK"
-    }
-
-private fun Purpose.toServerPurpose(): String =
-    when (this) {
-        Purpose.CONTENT_DELIVERY -> "INFO"
-        Purpose.IMPROVE_UNDERSTANDING -> "UNDERSTANDING"
-        Purpose.BUILD_EMPATHY -> "EMPATHY"
-    }
-
-private fun Style.toServerStyle(): String =
-    when (this) {
-        Style.PROFESSIONAL -> "FORMAL"
-        Style.FRIENDLY -> "FRIENDLY"
-        Style.CALM -> "CALM"
-        Style.COMFORTABLE -> "CASUAL"
-    }
-
-private fun Audience.toServerAudience(): String =
-    when (this) {
-        Audience.GENERAL_AUDIENCE -> "GENERAL"
-        Audience.EXPERT -> "PROFESSIONAL"
-        Audience.TEAMMATES -> "TEAMMATE"
-    }
-
-private fun String.toDomainCategory(): Category =
-    when (this) {
-        "OFFER" -> Category.PERSUASION
-        "EVENT" -> Category.EVENT
-        "EDUCATION" -> Category.EDUCATION
-        "WORK" -> Category.REPORT
-        else -> Category.REPORT
-    }
-
-private fun String.toDomainPurpose(): Purpose =
-    when (this) {
-        "INFO" -> Purpose.CONTENT_DELIVERY
-        "UNDERSTANDING" -> Purpose.IMPROVE_UNDERSTANDING
-        "EMPATHY" -> Purpose.BUILD_EMPATHY
-        else -> Purpose.CONTENT_DELIVERY
-    }
-
-private fun String.toDomainStyle(): Style =
-    when (this) {
-        "FORMAL" -> Style.PROFESSIONAL
-        "FRIENDLY" -> Style.FRIENDLY
-        "CALM" -> Style.CALM
-        "CASUAL" -> Style.COMFORTABLE
-        else -> Style.PROFESSIONAL
-    }
-
-private fun String.toDomainAudience(): Audience =
-    when (this) {
-        "GENERAL" -> Audience.GENERAL_AUDIENCE
-        "PROFESSIONAL" -> Audience.EXPERT
-        "TEAMMATE" -> Audience.TEAMMATES
-        else -> Audience.GENERAL_AUDIENCE
-    }
-
-private fun String.toAnalysisInstant(): Instant =
-    runCatching { Instant.parse(this) }
-        .getOrElse {
-            LocalDateTime
-                .parse(this)
-                .toInstant(TimeZone.of("Asia/Seoul"))
-        }
