+++ Prezel/core/domain/src/main/kotlin/com/team/prezel/core/domain/repository/presentation/PresentationRepository.kt
@@ -1,12 +1,25 @@
 package com.team.prezel.core.domain.repository.presentation
 
-import com.team.prezel.core.model.presentation.PresentationAnalysisRequest
+import com.team.prezel.core.model.presentation.Audience
+import com.team.prezel.core.model.presentation.Category
 import com.team.prezel.core.model.presentation.PresentationAnalysisSummary
 import com.team.prezel.core.model.presentation.PresentationScriptDetail
 import com.team.prezel.core.model.presentation.PresentationWordDetail
+import com.team.prezel.core.model.presentation.Purpose
+import com.team.prezel.core.model.presentation.Style
 
 interface PresentationRepository {
-    suspend fun analyzePresentation(request: PresentationAnalysisRequest): Result<PresentationAnalysisSummary>
+    suspend fun analyzePresentation(
+        name: String,
+        date: String,
+        category: Category,
+        purpose: Purpose,
+        style: Style,
+        audience: Audience,
+        script: String?,
+        scriptFilePath: String?,
+        audioFilePath: String,
+    ): Result<PresentationAnalysisSummary>
 
     suspend fun reAnalyzePresentation(
         presentationId: Long,
