+++ Prezel/core/network/src/main/java/com/team/prezel/core/network/datasource/PracticeRemoteDataSourceImpl.kt
@@ -2,7 +2,6 @@ package com.team.prezel.core.network.datasource
 
 import com.team.prezel.core.network.model.practice.AnalyzePracticeRecordingResponse
 import com.team.prezel.core.network.model.practice.PracticeSentenceResponse
-import com.team.prezel.core.network.model.practice.PresentationRecordingAnalysisResponse
 import com.team.prezel.core.network.model.requireData
 import com.team.prezel.core.network.service.PracticeService
 import io.ktor.client.request.forms.ChannelProvider
@@ -46,60 +45,8 @@ internal class PracticeRemoteDataSourceImpl @Inject constructor(
             ).requireData()
     }
 
-    override suspend fun analyzePresentationRecording(
-        name: String,
-        date: String,
-        type: String,
-        purpose: String,
-        style: String,
-        audience: String,
-        script: String?,
-        scriptFilePath: String?,
-        audioFilePath: String,
-    ): PresentationRecordingAnalysisResponse {
-        require(audioFilePath.isNotBlank()) { "발표 음성 파일 경로가 비어 있습니다." }
-
-        val audioFile = File(audioFilePath)
-        val scriptFile = scriptFilePath
-            ?.takeIf(String::isNotBlank)
-            ?.let(::File)
-        val multipart = MultiPartFormDataContent(
-            formData {
-                append("name", name)
-                append("date", date)
-                append("type", type)
-                append("purpose", purpose)
-                append("style", style)
-                append("audience", audience)
-                script?.takeIf(String::isNotBlank)?.let { append("script", it) }
-                scriptFile?.let { file ->
-                    append(
-                        key = "scriptFile",
-                        value = file.toChannelProvider(),
-                        headers = Headers.build {
-                            append(HttpHeaders.ContentType, "text/${file.extension}")
-                            append(HttpHeaders.ContentDisposition, "filename=\"${file.name}\"")
-                        },
-                    )
-                }
-                append(
-                    key = "audio",
-                    value = audioFile.toChannelProvider(),
-                    headers = Headers.build {
-                        append(HttpHeaders.ContentType, "audio/${audioFile.extension}")
-                        append(HttpHeaders.ContentDisposition, "filename=\"${audioFile.name}\"")
-                    },
-                )
-            },
-        )
-
-        return practiceService
-            .analyzePresentationRecording(multipart = multipart)
-            .requireData()
-    }
+    private fun File.toChannelProvider(): ChannelProvider =
+        ChannelProvider(size = length()) {
+            inputStream().toByteReadChannel()
+        }
 }
-
-private fun File.toChannelProvider(): ChannelProvider =
-    ChannelProvider(size = length()) {
-        inputStream().toByteReadChannel()
-    }
