+++ Prezel/core/network/src/main/java/com/team/prezel/core/network/datasource/PresentationRemoteDataSource.kt
@@ -1,12 +1,21 @@
 package com.team.prezel.core.network.datasource
 
-import com.team.prezel.core.network.model.presentation.PresentationAnalysisRequestParts
 import com.team.prezel.core.network.model.presentation.PresentationScriptDetailResponse
 import com.team.prezel.core.network.model.presentation.PresentationSummaryResponse
 import com.team.prezel.core.network.model.presentation.PresentationWordDetailResponse
 
 interface PresentationRemoteDataSource {
-    suspend fun analyzePresentation(request: PresentationAnalysisRequestParts): PresentationSummaryResponse
+    suspend fun analyzePresentation(
+        name: String,
+        date: String,
+        type: String,
+        purpose: String,
+        style: String,
+        audience: String,
+        script: String?,
+        scriptFilePath: String?,
+        audioFilePath: String,
+    ): PresentationSummaryResponse
 
     suspend fun reAnalyzePresentation(
         presentationId: Long,
