--- Prezel/core/network/src/main/java/com/team/prezel/core/network/model/presentation/PresentationAnalysisRequestParts.kt
@@ -1,12 +1,0 @@
-package com.team.prezel.core.network.model.presentation
-
-data class PresentationAnalysisRequestParts(
-    val name: String,
-    val date: String,
-    val type: String,
-    val purpose: String,
-    val style: String,
-    val audience: String,
-    val script: String?,
-    val audioFilePath: String,
-)
